+++ frontend/src/Cabinet/recoil/selectors.ts
@@ -6,16 +6,18 @@ import {
   currentFloorCabinetState,
   currentFloorNumberState,
   currentSectionNameState,
+  userState,
 } from "@/Cabinet/recoil/atoms";
+import { TOAuthProviderOrEmpty } from "@/Cabinet/components/Card/SocialAccountLinkCard/SocialAccountLinkCard.container";
 import {
   IFloorSectionColNum,
   ISectionColNum,
 } from "@/Cabinet/assets/data/sectionColNumData";
 import {
-  CabinetInfo,
   CabinetPreviewInfo,
   ICurrentSectionInfo,
 } from "@/Cabinet/types/dto/cabinet.dto";
+import { IUserOAuthLinkInfoDto } from "@/Cabinet/types/dto/oAuth.dto";
 
 export const buildingsState = selector<Array<string>>({
   key: "Buildings",
@@ -112,3 +114,26 @@ export const currentSectionColNumState = selector<number | undefined>({
     return currentFloorColNum[currentSectionIdx].colNum;
   },
 });
+
+export const linkedOAuthInfoState = selector<IUserOAuthLinkInfoDto | null>({
+  key: "LinkedOAuthInfo",
+  get: ({ get }) => {
+    const myInfo = get(userState);
+
+    return myInfo.userOauthConnection;
+  },
+});
+// TODO : 다른 데에서도 사용하는지 확인하고 적용
+
+export const linkedProviderState = selector<TOAuthProviderOrEmpty>({
+  key: "LinkedProvider",
+  get: ({ get }) => {
+    const linkedOAuthInfo = get(linkedOAuthInfoState);
+    const linkedProvider: TOAuthProviderOrEmpty = linkedOAuthInfo
+      ? linkedOAuthInfo.providerType
+      : "";
+
+    return linkedProvider;
+  },
+});
+// TODO : 다른 데에서도 사용하는지 확인하고 적용
